requirements.txt
streamlit>=1.28.0
openai>=1.50.0
Pillow>=10.0.0
requests>=2.31.0